refactor: split monolithic cli.py into focused command modules (#432)
Split the 2,045-line cli.py into 6 focused command modules under
ember/entrypoints/commands/. Moved _get_targeted_sync_hint to core
layer. Reduced all function complexity to B(9) or below (down from
C(12)-C(15)). All 1,691 existing tests pass unchanged.
