+++ ember/core/sync/__init__.py
@@ -4,6 +4,10 @@ Contains the SyncService for checking staleness and executing sync operations,
 with proper error classification.
 """
 
-from ember.core.sync.sync_service import SyncService, classify_sync_error
+from ember.core.sync.sync_service import (
+    SyncService,
+    classify_sync_error,
+    get_targeted_sync_hint,
+)
 
-__all__ = ["SyncService", "classify_sync_error"]
+__all__ = ["SyncService", "classify_sync_error", "get_targeted_sync_hint"]
