+++ ember/core/sync/sync_service.py
@@ -62,6 +62,51 @@ def classify_sync_error(exception: Exception) -> SyncErrorType:
     return SyncErrorType.UNKNOWN
 
 
+def get_targeted_sync_hint(error_message: str) -> str:
+    """Get a targeted hint based on the sync error message.
+
+    Analyzes the error message to provide specific, actionable guidance
+    rather than generic hints.
+
+    Args:
+        error_message: The error message from the failed sync operation.
+
+    Returns:
+        An actionable hint string specific to the error type.
+    """
+    error_lower = error_message.lower()
+
+    # Database corruption - suggest reindex
+    if any(
+        term in error_lower
+        for term in ["malformed", "corrupt", "database disk image", "not a database"]
+    ):
+        return "Run 'ember sync --reindex' to rebuild the index"
+
+    # Model mismatch or dimension mismatch - suggest init --force
+    if "model" in error_lower and ("changed" in error_lower or "mismatch" in error_lower):
+        return "Run 'ember init --force' to reinitialize with the new model"
+
+    # Vector dimension mismatch (model changed but message doesn't say "model")
+    if "dimension" in error_lower and "mismatch" in error_lower:
+        return "Run 'ember init --force' to reinitialize with the new model"
+
+    # Permission issues
+    if "permission" in error_lower or "access denied" in error_lower:
+        return "Check file permissions in .ember/ directory"
+
+    # Disk space issues
+    if "no space left" in error_lower or "disk full" in error_lower:
+        return "Free up disk space and retry"
+
+    # Lock/busy database
+    if "locked" in error_lower or "busy" in error_lower:
+        return "Another process may be using the database. Wait and retry."
+
+    # Generic fallback - suggest verbose mode
+    return "Run 'ember sync --verbose' for more details"
+
+
 class SyncService:
     """Service for coordinating index synchronization.
