+++ ember/entrypoints/commands/__init__.py
@@ -1,0 +1,4 @@
+"""Command modules for the Ember CLI.
+
+Each module registers its commands with the main CLI group.
+"""
